requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
mss>=6.0.0
Pillow>=9.0.0
